x/contract/types/message_call_contract_test.go
package types

import (
	"testing"

	"github.com/arran8901/chainlog-platform/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCallContract_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCallContract
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCallContract{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCallContract{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
